+++ src/views/ContractDataProvider.ts
@@ -48,7 +48,7 @@ export class ContractDataProvider implements vscode.TreeDataProvider<Contract> {
 			contract.description = contract.contractAddress;
 			contract.tooltip = getContractTooltip();
 			contract.contextValue = Constants.VIEWS_CONTRACT;
-			contract.iconPath = contract.chainConfig == Workspace.GetWorkspaceChainConfig().configName ?  new vscode.ThemeIcon("plug"): "";
+			contract.iconPath = contract.chainConfig == Workspace.GetWorkspaceChainConfig().configName ?  "" : new vscode.ThemeIcon("debug-disconnect");
 			contract.command = {
 				title: "Select Contract",
 				command: "cosmy-wasmy.selectContract",
@@ -62,6 +62,10 @@ export class ContractDataProvider implements vscode.TreeDataProvider<Contract> {
 			tooltip += "Creator: " + contract.creator;
 			if(contract.notes && contract.notes.trim().length > 0) {
 				tooltip += "\n\n"  + contract.notes;
+			}
+			if(contract.chainConfig != Workspace.GetWorkspaceChainConfig().configName) {
+				tooltip += "\n\n";
+				tooltip += "$(alert) *The imported contracts are not associated with any of the configured chains. Delete and reimport the contract to fix this.*";		
 			}
 			return new vscode.MarkdownString(tooltip, true);
 		}